+++ app/package.json
@@ -14,7 +14,8 @@
     "normalize.css": "^8.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.9.0"
+    "react-router-dom": "^6.9.0",
+    "vite-plugin-eslint": "^1.8.1"
   },
   "devDependencies": {
     "@types/node": "^18.15.1",
