+++ app/src/components/pages/Home.tsx
@@ -1,5 +1,5 @@
 import { Component } from 'react';
-import { SearchBar } from '../SearchBar';
+import { SearchBar } from '../SearchBar/SearchBar';
 
 export class Home extends Component {
   render() {
